OH MY GOD WE CAN HAVE INTEGERS

+++ game/src/api/Player.hpp
@@ -10,8 +10,8 @@ void fromString(const std::string &str);
         bool connected;
         std::shared_ptr<bool> crowned;
         std::shared_ptr<std::string> name;
-        double score;
+        int64_t score;
         std::shared_ptr<std::string> user_id;
-        double win_count;
+        int64_t win_count;
     };
 }
