+++ game/src/api/Room.hpp
@@ -10,6 +10,6 @@ void fromString(const std::string &str);
         std::string code;
         std::string host_name;
         std::string last_updated;
-        double player_count;
+        int64_t player_count;
     };
 }
