+++ game/src/api/UpdateMsg.hpp
@@ -11,9 +11,9 @@ namespace API {
     struct UpdateMsg {
 std::string toString() const;
 void fromString(const std::string &str);
-        double id;
+        int64_t id;
         std::shared_ptr<bool> reset;
-        double score;
+        int64_t score;
         UpdateMsgType type = static_cast<UpdateMsgType>(0);
         std::shared_ptr<std::vector<bool>> used;
     };
